requirements.txt
# Core Django
Django==5.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2   # needed on Windows

# REST API
djangorestframework==3.16.0

# HTTP client to call LLaMA / Ollama
httpx==0.27.2
